Fixes ItemRss2Test so that it shows test errors instead of giving fatal errors when no files were inserted.

+++ application/tests/integration/Models/ItemRss2Test.php
@@ -32,6 +32,14 @@ class Omeka_Model_ItemRss2Test extends Omeka_Test_AppTestCase
         $_SERVER['HTTP_HOST'] = 'localhost';
     }
     
+    public function assertPreConditions()
+    {
+        $this->assertEquals(1, $this->db->getTable('Item')->count(),
+            "There should be one item in the database.");
+        $this->assertEquals(1, $this->db->getTable('File')->count(),
+            "There should be one file in the database.");
+    }
+    
     public function testCanGetValidItemRss2Output()
     {   
         $this->dispatch('items/browse?output=rss2');
@@ -65,7 +73,9 @@ class Omeka_Model_ItemRss2Test extends Omeka_Test_AppTestCase
     {
         // Delete the physical files that were ingested in setUp().
         $testFile = $this->db->getTable('File')->find(1);
-        $testFile->delete();
+        if ($testFile instanceof File) {
+            $testFile->delete();
+        }
         parent::tearDown();
     }
 }
